Initial step for unification
- Base classes for both device and contact.
- Separating out shared functionality.
- QTBM still working while SCBA not using new things.
