+++ src/quatrex/pre_processsing/pre_process.py
@@ -4,7 +4,7 @@
 
 from qttools.comm import comm
 from quatrex.core.config import QuatrexConfig
-from quatrex.device import Device
+from quatrex.device import create_device
 from quatrex.pre_processsing.contact_bandstructure import plot_contact_band_structure
 from quatrex.pre_processsing.contact_fermi_level import pre_process_fermi_level
 
@@ -30,7 +30,7 @@ def pre_process(config: QuatrexConfig):
 
     device = None
     if config.formalism == "wf":
-        device = Device(config)
+        device = create_device(config)
 
     if config.pre_process.compute_fermi_level:
         pre_process_fermi_level(config, device)
