+++ tests/quatrex/bandstructure/test_band_edges.py
@@ -8,7 +8,7 @@ import pytest
 from qttools import xp
 from quatrex.bandstructure.band_edges import _compute_eigenvalues
 from quatrex.core.config import QuatrexConfig, parse_config, setup_context
-from quatrex.device import Device
+from quatrex.device import BaseDevice
 from quatrex.device.inputs import assemble_matrix, get_block_sizes
 from quatrex.grid import get_electron_energies
 
@@ -38,7 +38,7 @@ def _intialize(config: QuatrexConfig):
 
     potential = xp.zeros(hamiltonian.shape[-1], dtype=hamiltonian.dtype)
 
-    grid, __, __, __ = Device.load_structure(config)
+    grid, __, __, __ = BaseDevice._load_structure(config)
     block_sizes = get_block_sizes(config, grid)
     energies = get_electron_energies(config)
     kpoint_grid = config.device.kpoint_grid
